+++ build.ps1
@@ -1,14 +1,51 @@
+# spell-checker:ignore nologo psake
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute(
+  'PSReviewUnusedParameter',
+  'Command',
+  Justification = 'false positive'
+)]
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute(
+  'PSReviewUnusedParameter',
+  'Parameter',
+  Justification = 'false positive'
+)]
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute(
+  'PSReviewUnusedParameter',
+  'CommandAst',
+  Justification = 'false positive'
+)]
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute(
+  'PSReviewUnusedParameter',
+  'FakeBoundParams',
+  Justification = 'false positive'
+)]
+[Diagnostics.CodeAnalysis.SuppressMessageAttribute(
+  'PSReviewUnusedParameter',
+  'Help',
+  Justification = 'false positive'
+)]
 [CmdletBinding(DefaultParameterSetName = 'Task')]
 param(
   # Build task(s) to execute
-  [parameter(ParameterSetName = 'task', position = 0)]
+  [Parameter(ParameterSetName = 'task', Position = 0)]
+  [ArgumentCompleter( {
+      param($Command, $Parameter, $WordToComplete, $CommandAst, $FakeBoundParams)
+      try {
+        Get-PSakeScriptTasks -BuildFile './psakefile.ps1' -ErrorAction 'Stop' |
+          Where-Object { $_.Name -like "$WordToComplete*" } |
+          Select-Object -ExpandProperty 'Name'
+      } catch {
+        # Silently fail if psake tasks can't be retrieved
+        @()
+      }
+    })]
   [string[]]$Task = 'default',
 
   # Bootstrap dependencies
   [switch]$Bootstrap,
 
   # List available build tasks
-  [parameter(ParameterSetName = 'Help')]
+  [Parameter(ParameterSetName = 'Help')]
   [switch]$Help,
 
   # Optional properties to pass to psake